Updated demangler on Linux.

+++ src/roo_logging/demangle.h
@@ -78,6 +78,6 @@ namespace roo_logging {
 // Demangle "mangled".  On success, return true and write the
 // demangled symbol name to "out".  Otherwise, return false.
 // "out" is modified even if demangling is unsuccessful.
-bool Demangle(const char *mangled, char *out, int out_size);
+bool Demangle(const char *mangled, char *out, size_t out_size);
 
 }  // namespace roo_logging
